Timer adjusted for real seconds Timer ending turn, going through after-turn states correctly, then switching to next team and returning to WormSelect. The games main life cycle with the state machine seems to be working.
There seems to be some kind of error with the sudden death timer, as changing code that involves that variable will sometimes yield builds that crash on start.

For now, TurnEnd, Death, AfterTurn all take 2 seconds before auto-changing. In the future, these would wait for the games events to finish instead of arbitrary 2 seconds.